Перед удалением закомен. кода
